Adding additional semaphores for each swapchain image; Initialize cpu fence At signal start instead of unsignaled state, prevents having to use a branching statement to check if gpu is one frame behind.

+++ Engine/Renderer/Public/Renderer/RHI/Swapchain.hpp
@@ -32,8 +32,8 @@ public:
 
   VkSwapchainKHR                Handle() { return mSwapchain; }
 
-
-  size_t                        ImageCount() const { return SwapchainImages.size(); }
+  // Current number of images available in the swapchain.
+  u32                        ImageCount() const { return static_cast<u32>(SwapchainImages.size()); }
 
   SwapchainImage&               Get(const size_t index) { return SwapchainImages[index]; }
   SwapchainImage&               operator[](const size_t index) { return SwapchainImages[index]; }
